# docs/api-docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: An's DBMS Database Object API
  version: 1.0.0
  description: |
    REST API for managing the main database objects in An's DBMS.

    Supported resources:

    * Database
    * Schema
    * Table
    * Column
    * View
    * Index
    * Constraint

    Resource hierarchy:

    Database
      └── Schema
          ├── Table
          │   ├── Column
          │   ├── Constraint
          │   └── Index
          └── View

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /databases:
    post:
      tags:
        - Databases
      summary: Create a database
      operationId: createDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatabaseRequest'
      responses:
        '201':
          description: Database created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Databases
      summary: List all databases
      operationId: listDatabases
      responses:
        '200':
          description: Databases retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatabaseResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /databases/{databaseId}:
    parameters:
      - $ref: '#/components/parameters/DatabaseId'

    get:
      tags:
        - Databases
      summary: Get a database by ID
      operationId: getDatabase
      responses:
        '200':
          description: Database retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Databases
      summary: Delete a database
      operationId: deleteDatabase
      responses:
        '204':
          description: Database deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /databases/{databaseId}/name:
    parameters:
      - $ref: '#/components/parameters/DatabaseId'

    patch:
      tags:
        - Databases
      summary: Rename a database
      operationId: renameDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameDatabaseRequest'
      responses:
        '200':
          description: Database renamed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /databases/{databaseId}/close:
    parameters:
      - $ref: '#/components/parameters/DatabaseId'

    post:
      tags:
        - Databases
      summary: Close a database
      operationId: closeDatabase
      responses:
        '200':
          description: Database closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /databases/{databaseId}/read-only:
    parameters:
      - $ref: '#/components/parameters/DatabaseId'

    patch:
      tags:
        - Databases
      summary: Change the database read-only mode
      operationId: setDatabaseReadOnly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetReadOnlyRequest'
      responses:
        '200':
          description: Read-only mode updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # =========================================================
  # Schema
  # =========================================================
...
components:
...
  parameters:
    DatabaseId:
      name: databaseId
      in: path
      required: true
      description: Database identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # =========================================================
  # Schemas
  # =========================================================
  schemas:
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Schema was not found
        path:
          type: string
          example: /api/v1/schemas/123
        validationErrors:
          type: array
          items:
            type: string

    # Database DTOs
    CreateDatabaseRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: customer_database
    RenameDatabaseRequest:
      type: object
      required:
        - newName
      properties:
        newName:
          type: string
          minLength: 1
          example: customer_management
    SetReadOnlyRequest:
      type: object
      required:
        - readOnly
      properties:
        readOnly:
          type: boolean
          example: true
    DatabaseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        state:
          $ref: '#/components/schemas/DatabaseStateType'
        readOnly:
          type: boolean
        schemaCount:
          type: integer
          minimum: 0
    DatabaseStateType:
      type: string
      enum:
        - CLOSED
        - OPEN
        - READ_ONLY

    # Schema DTOs
...